refactor: unit test
