+++ src/components/Card/index.tsx
@@ -1,8 +1,8 @@
 import { FC } from 'react'
-import style from './card.css'
+// import style from './card.css'
 
 const Card: FC<{ title: string }> = ({ title, children }) => (
-  <div className={style.Card}>
+  <div>
     <h4>{title}</h4>
     {children}
   </div>
