+++ src/pages/home.test.tsx
@@ -1,5 +1,10 @@
 import { render } from '@testing-library/react'
+import Home from './Home'
 describe('HomePage', () => {
-  it.todo('displays last 3 space X launchs')
+  it.skip('displays without launches', () => {
+    const { getByRole, debug } = render(<Home data={{ data: { launchesPast: [] } }} />)
+    getByRole('heading')
+    debug()
+  })
   it.todo('has a link to detail page')
 })
